Move TargetFactory interface to the top-level namespace
