+++ src/Converter/GenericConverter.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Neusta\ConverterBundle\Converter;
 
 use Neusta\ConverterBundle\Converter;
-use Neusta\ConverterBundle\Target\TargetFactory;
+use Neusta\ConverterBundle\TargetFactory;
 use Neusta\ConverterBundle\Populator;
 
 /**
