+++ tests/app/config/packages/neusta_converter.yaml
@@ -2,7 +2,7 @@ neusta_converter:
   converter:
     test.person.converter:
 #      converter: Neusta\ConverterBundle\Converter\GenericConverter
-      target_factory: Neusta\ConverterBundle\Tests\Fixtures\Target\PersonFactory
+      target_factory: Neusta\ConverterBundle\Tests\Fixtures\Model\PersonFactory
       populators:
         - Neusta\ConverterBundle\Tests\Fixtures\Populator\PersonNamePopulator
 #      properties:
@@ -11,7 +11,7 @@ neusta_converter:
 
     test.person.converter.with.cache:
 #      converter: Neusta\ConverterBundle\Converter\GenericConverter
-      target_factory: Neusta\ConverterBundle\Tests\Fixtures\Factory\PersonFactory
+      target_factory: Neusta\ConverterBundle\Tests\Fixtures\Model\PersonFactory
       populators:
         - Neusta\ConverterBundle\Tests\Fixtures\Populator\PersonNamePopulator
       cached:
